fixed error messages and conversion types

+++ print_nums.c
@@ -52,7 +52,7 @@ int handle_oct(va_list oct)
 {
 	char tmp[100];
 
-	convert_bases(va_arg(oct, unsigned long), tmp, 8, 0);
+	convert_bases(va_arg(oct, unsigned int), tmp, 8, 0);
 
 	return (_puts(tmp));
 }
@@ -68,7 +68,7 @@ int handle_b(va_list b)
 {
 	char tmp[100];
 
-	convert_bases(va_arg(b, unsigned long), tmp, 2, 0);
+	convert_bases(va_arg(b, unsigned int), tmp, 2, 0);
 
 	return (_puts(tmp));
 }
@@ -84,7 +84,7 @@ int handle_u(va_list u)
 {
 	char tmp[100];
 
-	convert_bases(va_arg(u, unsigned long), tmp, 10, 0);
+	convert_bases(va_arg(u, unsigned int), tmp, 10, 0);
 
 	return (_puts(tmp));
 }
